move card and list item to sandbox component
