+++ packages/app/src/app/pages/NewDashboard/Components/Sandbox/index.tsx
@@ -4,8 +4,8 @@ import { sandboxUrl } from '@codesandbox/common/lib/utils/url-generator';
 import { ESC } from '@codesandbox/common/lib/utils/keycodes';
 import { Element } from '@codesandbox/components';
 import { useOvermind } from 'app/overmind';
-import { SandboxItem } from '../SandboxItem';
-import { SandboxCard } from '../SandboxCard';
+import { SandboxCard } from './SandboxCard';
+import { SandboxListItem } from './SandboxListItem';
 
 export const Sandbox = ({ sandbox, isTemplate = false, ...props }) => {
   const {
@@ -77,7 +77,7 @@ export const Sandbox = ({ sandbox, isTemplate = false, ...props }) => {
   if (dashboard.viewMode === 'list' || location.pathname.includes('deleted')) {
     return (
       <Element style={{ gridColumn: '1/-1' }}>
-        <SandboxItem {...sandboxProps} {...props} />
+        <SandboxListItem {...sandboxProps} {...props} />
       </Element>
     );
   }
